chore: update nextjs example to use ix-player
